% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_scz_2020}
\alias{covid_scz_2020}
\title{Daily new COVID-19 cases, Santa Cruz County, California}
\format{
A numeric vector of length 518.
}
\source{
California Department of Public Health,
\url{https://lab.data.ca.gov/dataset/covid-19-time-series-metrics-by-county-and-state-archived}.
}
\usage{
covid_scz_2020
}
\description{
Daily counts of newly reported COVID-19 cases in Santa Cruz
County, California from July 1, 2020 through December 1, 2021.
}
\keyword{datasets}
