% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp-lowlevel-docs.R
\name{dgtf-cpp-lowlevel}
\alias{dgtf-cpp-lowlevel}
\alias{dgtf_default_algo_settings}
\alias{dgtf_default_model}
\alias{dgtf_forecast}
\alias{dgtf_infer}
\alias{dgtf_posterior_predictive}
\alias{dgtf_simulate}
\alias{dlognorm0}
\alias{dnbinom0}
\alias{lognorm2serial}
\alias{serial2lognorm}
\title{Low-level C++ interface functions}
\description{
These functions are generated from the package's Rcpp interface and are
exported for advanced use and for the higher-level R wrappers. Most users
should call \code{dgtf()}, \code{dgtf_model()}, \code{posterior_predict()},
or \code{dgtf_forecast_fit()} instead.
}
\keyword{internal}
