% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dgtf-gain}
\alias{dgtf-gain}
\alias{gain_identity}
\alias{gain_softplus}
\alias{gain_ramp}
\alias{gain_exponential}
\alias{gain_logistic}
\title{Gain-function components}
\usage{
gain_identity()

gain_softplus()

gain_ramp()

gain_exponential()

gain_logistic()
}
\value{
A \code{dgtf_gain} component object.
}
\description{
Differentiable transforms applied to the first state element inside the
transfer function.
}
