% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dgtf-link}
\alias{dgtf-link}
\alias{link_identity}
\alias{link_exponential}
\alias{link_logistic}
\title{Link-function components}
\usage{
link_identity()

link_exponential()

link_logistic()
}
\value{
A \code{dgtf_link} component object.
}
\description{
Maps the conditional mean to a linear predictor.
}
