% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgtf-package.R
\docType{package}
\name{dgtf-package}
\alias{dgtf-package}
\title{dgtf: Variational and Monte Carlo Inference for Bayesian State-Space Models of Count Data}
\description{
The \code{dgtf} package provides a unified interface to several inference
algorithms (variational Bayes, MCMC, sequential Monte Carlo, and linear
Bayes) for a general class of dynamic generalised time-series filters
for count data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/minimeini/dgtf}
  \item Report bugs at \url{https://github.com/minimeini/dgtf/issues}
}

}
\author{
\strong{Maintainer}: Meini Tang \email{mtang41@ucsc.edu} (\href{https://orcid.org/0000-0001-5338-8772}{ORCID})

Authors:
\itemize{
  \item Meini Tang \email{mtang41@ucsc.edu} (\href{https://orcid.org/0000-0001-5338-8772}{ORCID})
  \item Raquel Prado \email{rprado@ucsc.edu} (\href{https://orcid.org/0000-0001-6709-7191}{ORCID})
}

}
\keyword{internal}
