% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{lba_control}
\alias{lba_control}
\title{Linear-Bayes control}
\usage{
lba_control(discount_factor = 0.95, ...)
}
\arguments{
\item{discount_factor}{Discount factor in (0, 1].}

\item{...}{Forwarded to the C++ engine.}
}
\value{
A \code{dgtf_lba_control} object.
}
\description{
Tuning parameters for the linear-Bayes approximation
(\code{method = "linear_bayes"}).
}
