% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{mcmc_control}
\alias{mcmc_control}
\title{MCMC control}
\usage{
mcmc_control(iter = 2000L, warmup = 1000L, thin = 1L, ...)
}
\arguments{
\item{iter}{Number of MCMC iterations (post-warmup retained).}

\item{warmup}{Burn-in iterations, discarded.}

\item{thin}{Thinning interval.}

\item{...}{Forwarded to the C++ engine.}
}
\value{
A \code{dgtf_mcmc_control} object.
}
\description{
Tuning parameters for the disturbance-MCMC sampler
(\code{method = "mcmc"}).
}
